factor out innerlist/params/bare_item
